Overlay Tunnel Endpoint Manual Setting by CIDR
This feature enhancement enables the user to set the desired
outgoing interface and its tunnel endpoint IP(v4) address by
a designated CIDR, which may be different from the default
which currently is just the Kubernetes node IP.

The usage of the this feature is just by adding the following
declaration in the installation yaml file:

apiVersion: v1
kind: ConfigMap
metadata:
  name: cilium-config
  namespace: kube-system
data:
  ...
  tunnel-endpoint-ipv4-cidr: "172.168.50.0/24"

This feature supports VXLAN overlay tunnel, which would lead the
data traffic and tunnel selection from the default k8s node IP
and related interface to the IP address which matches the setted
IPv4 CIDR and its interface.

If this feature is enabled by setting the desired CIDR,
the MTU and NodePort (physical) device would also be changed to this
matching interface(device) to enable the working of BPF on it.

This feature currently doesn't support IPSEC or node encryption.
This feature is still in its initial state currently which just
support 1 CIDR specified, and the multi-CIDRs case would be
supported if needed.

For further information, please look for:
https://github.com/cilium/cilium/issues/14520

Signed-off-by: trevor tao <trevor.tao@arm.com>
Signed-off-by: TrevorTaoARM <trevor.tao@arm.com>